move the global tracer provider code into the trace directory
