+++ api/include/opentelemetry/trace/global_provider.h
