--- api/test/global/BUILD
@@ -1,10 +1,0 @@
-cc_test(
-      name = "global",
-      srcs = [
-          "global_test.cc",
-      ],
-    deps = [
-        "//api",
-        "@com_google_googletest//:gtest_main",
-    ],
-)
